+++ src/Enqueuer.Queueing.API/Application/Commands/RenameQueueCommand.cs
@@ -4,13 +4,13 @@ namespace Enqueuer.Queueing.API.Application.Commands;
 
 public class RenameQueueCommand : IRequest
 {
-    public RenameQueueCommand(int id, string newQueueName)
+    public RenameQueueCommand(long id, string newQueueName)
     {
         Id = id;
         NewQueueName = newQueueName;
     }
 
-    public int Id { get; }
+    public long Id { get; }
 
     public string NewQueueName { get; }
 }
