+++ src/Enqueuer.Queueing.API/Application/Queries/GetQueueQuery.cs
@@ -5,10 +5,10 @@ namespace Enqueuer.Queueing.API.Application.Queries;
 
 public class GetQueueQuery : IRequest<Queue>
 {
-    public GetQueueQuery(int id)
+    public GetQueueQuery(long id)
     {
         Id = id;
     }
 
-    public int Id { get; }
+    public long Id { get; }
 }
